cleanup, moved api to handler/admin/api package
